requirement_molmosam.txt
accelerate==1.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.4
attrs==24.3.0
Babel==2.11.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.2.0
Brotli==1.0.9
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.1
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
einops==0.8.0
eva-decord==0.6.1
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.115.6
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.2
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.10.0.84
overrides==7.4.0
packaging==24.2
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.8.0
pillow==10.2.0
pip==24.2
platformdirs==3.10.0
portalocker==3.1.1
prometheus_client==0.21.0
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.15.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
safetensors==0.5.1
SAM-2==1.0
scipy==1.15.0
Send2Trash==1.8.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
starlette==0.41.3
supervision==0.25.1
sympy==1.13.1
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
tool_server==0.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.1
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
